// src/shared/enums/socket-events.enum.ts
export enum SocketEventsEnum {
  BET = 'bet',
  BET_ERROR = 'bet:err',
  ROUND_JACKPOT_VALUES = 'jackpot-values',
  BET_SUCCESS = 'bet:success',
  ROUND_CLOSE = 'round:close',
  ROUND_END = 'round:end',
  ROUND_BET_TIME = 'round:end-bet-time',
  ROUND_START = 'round:start',
  WINNER = 'winner',
  WINNER_ERR = 'winner:err',
}